requirements.txt
customtkinter==5.1.2
firebase_admin==6.1.0
numpy==1.24.2
Pillow==9.4.0
phonenumbers==8.13.7
opencv-python~=4.7.0.72
requests~=2.28.1
imageio~=2.26.1